Refactored HR dashboard screen
